// dpbloom-client/src/app/core/api/model/globalUserDashboardDto.ts
/**
 * DPBloom | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BloomLevelPerformanceDto } from './bloomLevelPerformanceDto';


export interface GlobalUserDashboardDto { 
    averageCourseScore?: number;
    averageExamCompletion?: number;
    bloomPerformance?: Array<BloomLevelPerformanceDto>;
}